==> code/03_calculating_substitution_rates.Rmd <==
---
title: "Supplementary Text S3: Calculating substitution rates for selected Ensembl genomes"
author:
- name: Fabricio Almeida-Silva
  affiliation: |
    VIB-UGent Center for Plant Systems Biology, Ghent University, 
    Ghent, Belgium
- name: Yves Van de Peer
  affiliation: |
    VIB-UGent Center for Plant Systems Biology, Ghent University, 
    Ghent, Belgium
output:
  BiocStyle::pdf_document
vignette: |
  %\VignetteIndexEntry{Supplementary Text S3}
  %\VignetteEngine{knitr::rmarkdown}
  %\VignetteEncoding{UTF-8}
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(
    echo = TRUE,
    warning = FALSE, 
    eval = FALSE
)

Sys.setlocale("LC_TIME", "en_US.UTF-8")
Sys.setlocale("LC_PAPER", "en_US.UTF-8")
Sys.setlocale("LC_MEASUREMENT", "en_US.UTF-8")
```

# Introduction

Here, we will describe the code to calculate substitution rates for selected
genomes in Ensembl and Ensembl Genomes instances using the Bioconductor package 
`r BiocStyle::Biocpkg("doubletrouble")`. 

```{r load_packages, eval = TRUE}
library(syntenet)
library(doubletrouble)
library(here)
library(tidyverse)
library(BiocParallel)

set.seed(123) # for reproducibility
options(timeout = 1e10) # to allow download of big files

# Load helper functions
source(here("code", "utils.R"))
```

# Data loading

Here, we will load the data frames of species metadata and the lists of
duplicated gene pairs for each Ensembl instance.

```{r eval = TRUE}
# Load metadata
load(here("products", "result_files", "metadata_all.rda"))
names(metadata_all)

# Load duplicates
load(here("products", "result_files", "fungi_duplicates.rda"))
load(here("products", "result_files", "plants_duplicates.rda"))
```

# Calculating substitution rates

Next, we will calculate substitution rates ($K_a$, $K_s$, and $K_a/K_s$) for
duplicate pairs in all selected species, namely:

1. Three fungi species (*Saccharomyces cerevisiae*, *Candida glabrata*,
and *Schizosaccharomyces pombe*).

2. Four legume species (*Glycine max*, *Phaseolus vulgaris*, *Vitis vinifera*,
*Selaginella moellendorffii*).

```{r}
# Fungi - S. cerevisiae, Candida glabrata, and Schizosaccharomyces pombe
## Download CDS
selected_fungi <- c(
    "saccharomyces_cerevisiae", "candida_glabrata", "schizosaccharomyces_pombe"
)
fungi_cds <- get_cds_ensembl(selected_fungi, ensembl = "fungi")

## Calculate substitution rates
fungi_kaks <- pairs2kaks(
    gene_pairs_list = fungi_duplicates[selected_fungi],
    cds = fungi_cds,
    bp_param = BiocParallel::SnowParam(workers = 8)
)

# Plants - Glycine max, Phaseolus vulgaris, Vitis vinifera, and Selaginella moellendorffii
## Download CDS
selected_plants <- c(
    "glycine_max", "phaseolus_vulgaris", "vitis_vinifera", 
    "selaginella_moellendorffii"
)

plants_cds <- get_cds_ensembl(selected_plants, ensembl = "plants")

## Calculate substitution rates
plants_duplicates <- plants_duplicates[selected_plants]

plants_kaks <- pairs2kaks(
    gene_pairs_list = plants_duplicates,
    cds = plants_cds,
    bp_param = BiocParallel::SnowParam(workers = 8)
)
```

Saving objects as .rda files:

```{r}
save(
    fungi_kaks, compress = "xz",
    file = here("products", "result_files", "fungi_kaks.rda")
)

save(
    plants_kaks, compress = "xz",
    file = here("products", "result_files", "plants_kaks.rda")
)
```


# Session info {.unnumbered}

This document was created under the following conditions:

```{r sessionInfo, echo = FALSE, eval = TRUE}
sessioninfo::session_info()
```
